requirements.txt
transformers==4.53.1
torch==2.7.1
datasets==4.0.0
scikit-learn==1.7.0
sentence-transformers==5.0.0
PyMuPDF==1.26.3
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
httplib2==0.22.0
requests==2.32.4
python-dotenv==1.1.1
pandas==2.3.0
tqdm==4.67.1
